Improve CSS completion
Disabled autocompletion inside hex colors.
CSS variables are now wrapped with "var()" when completed as a value or postfixed with ": ;" when completed as a property.
CSS variable autocompletion is now present inside style attributes as well.
Fixed value completions being returned at the start of style attributes if an earlier attribute had a colon inside it.
Disabled implicit autocompletion when writing numbers.